web/requirements.txt
Flask
gunicorn
python-dotenv
redis
bcrypt
pytz